Use defined exit codes so STATE_OK isn't used unintentionally
utils.sh[1] doesn't define a STATE_ERROR. Use STATE_CRITICAL instead.
Without this, the plugin call `exit` with no arguments, which uses the
exit code of the previous command. Since the previous command is `echo`
in both cases, the script will exit with the 0 code, which is STATE_OK.
That means nagios will consider it successful when the service has
failed.

1. https://github.com/nagios-plugins/nagios-plugins/blob/master/plugins-scripts/utils.sh.in